requirements.txt
Pillow==9.5.0
PyGetWindow==0.0.9
pywin32==304
pywinauto==0.6.8
requests==2.28.2
Send2Trash==1.8.0